Fix bug in mongoseed.sh
The script was breaking if there were no elements in the `database/seed/` directory. In that case, it would try to read from a file literally named `database/seed/*.json`, which of course does not exist.

This commit fixes that error. Now, for each file in `database/seed/*.json`, the script will check to see if the file exists first. This catches the case where the script literally tries to seed `*.json`.