Change theme colors
The website now uses dark green colors with sea-green accents and orange
warnings. (Dark mode is similar except that the greens and oranges are
all pastel and desaturated.)

As a team, we had mostly settled on green as a primary color, but we
didn't know what we wanted our accent colors to be. In using green
accents, this commit tentatively picks one of the ideas that had been
floated and that nobody seemed to object strongly to. (The orange
warnings were a spur-of-the-moment pick.) If anyone doesn't like this
color scheme, that's all right--we can always find something different.