Get the URL from environment.ts
Okay, I know that Audrey feels strongly about not moving the URL into a separate file, and just inlining it into pdf.service.ts.

However:
Moving the URL into environment.ts keep all the functionality related to deployment in one place. When we deploy, we know that we need to mess with environment.ts, but we might never think to mess with pdf.service.ts. (In general, I feel like it's a bad idea if we have to change application code just because we're deploying.)

(Also, Prof. KK recommended putting it in environment.ts when I talked to her about it.)

If you all feel like this is a bad idea, though, I'm open to doing it in pdf.service.ts instead 😊

Addresses #10.