Add type RealJsPDF to expose more methods of jsPDF to TypeScript
The module @types/jspdf tells TypeScript what methods objects of class jsPDF have. However, it is awfully conservative with which methods it chooses to expose. Rather than expose all of jspdf's public API, it chooses only to expose those public API methods that are documented. Unfortunately, the documentation for jspdf is incomplete; quite a few public methods are undocumented, and hence unexposed by @types/jspdf.

This commit declares a new type, RealJsPDF, that exposes more public methods of the jsPDF class. (Not all of them—just the ones we need right now.)

Addresses #10.