Write e2e test for downloading the pdf
This test is a little weird because it messes with the filesystem on your computer.

Specifically, it looks in your downloads folder, removes DoorBoard.pdf if it exists, clicks the GENERATE PDF button, and then checks your downloads folder again to make sure that DoorBoard.pdf exists.

I've only ran it on a Mac so far—it *should* work on Windows and Linux, but that still needs to be verified.

Addresses #10.