Move RealJsPDF type from application code into test code
This commit removes RealJsPDF from the application code and moves it into the pdf.service.spec.ts file exclusively.

The RealJsPDF type exposes functionality that is only needed in test code. It's also a bit of a messy hack. As a rule, that sort of messiness should be confined to as narrow a scope as possible. So, as of now, we're only using it in test code.