Apply changes to NoteController
This should make it so that the only time that Auth0 is called is when a new owner is created. Whenever you do anything to a note, there will not be any calls to Auth0. This will prevent the server from throwing errors after making multiple requests within a couple of seconds.